Fixes and improvements

+++ src/main/java/com/fmalessio/twitterapi/service/BoardsServiceImpl.java
@@ -7,7 +7,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.fmalessio.twitterapi.entity.Board;
-import com.fmalessio.twitterapi.entity.Interest;
 import com.fmalessio.twitterapi.entity.SearchedTweet;
 import com.fmalessio.twitterapi.repository.BoardRepository;
 import com.fmalessio.twitterapi.repository.SearchedTweetRepository;
@@ -18,14 +17,11 @@ public class BoardsServiceImpl implements BoardsService {
 	private BoardRepository boardRepository;
 	private SearchedTweetRepository searchedTweetRepository;
 
-	private InterestsService interestsService;
-
 	@Autowired
 	public BoardsServiceImpl(BoardRepository boardRepo, SearchedTweetRepository searchedTweetRepository,
 			InterestsService interestsService) {
 		this.boardRepository = boardRepo;
 		this.searchedTweetRepository = searchedTweetRepository;
-		this.interestsService = interestsService;
 	}
 
 	@Override
@@ -42,12 +38,7 @@ public class BoardsServiceImpl implements BoardsService {
 	@Override
 	public void delete(long id) {
 		Optional<Board> board = boardRepository.findById(id);
-		if (board.isPresent()) {
-			List<Interest> interests = board.get().getInterests();
-			// It's necessary to remove all Quartz Jobs
-			for (Interest interest : interests) {
-				interestsService.delete(interest.getId());
-			}
+		if (board.isPresent() && board.get().getInterests().isEmpty()) {
 			boardRepository.deleteById(id);
 		}
 	}
